+++ src/main/java/net/ccbluex/liquidbounce/injection/mixins/blaze3d/MixinRenderSystem.java
@@ -24,6 +24,7 @@ import com.mojang.blaze3d.platform.Window;
 import com.mojang.blaze3d.systems.RenderSystem;
 import net.ccbluex.liquidbounce.render.ClientTesselator;
 import net.ccbluex.liquidbounce.render.GrowableMappableRingBuffer;
+import net.ccbluex.liquidbounce.render.utils.RenderingDebug;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
@@ -34,6 +35,7 @@ public abstract class MixinRenderSystem {
 
     @Inject(method = "flipFrame", at = @At("RETURN"))
     private static void onFlipFrame(Window window, TracyFrameCapture tracyFrameCapture, CallbackInfo ci) {
+        RenderingDebug.flipFrame();
         ClientTesselator.Shared.clear();
         GrowableMappableRingBuffer.cleanup();
     }
